requirements.txt
# see pipdeptree
lxml>=4.6.1
pbr>=5.5.1
confuse>=1.3.0
pyyaml>=5.3.1
schedule>=1.0.0
